Refactor Banner component structure and add CSS module for styling
