+++ frontend/src/Pages/HomePage.jsx
@@ -1,4 +1,4 @@
-import Banner from "../components/Banner"
+import Banner from "../components/Banner/Banner"
 
 function HomePage() {
     return (
