+++ frontend/src/components/Banner/Banner.module.css
@@ -1,0 +1,7 @@
+.banner{
+    background-color: var(--color-secondary);    
+}
+
+.registration{
+    background-color: var(--color-primary);
+}
